Updated remove holds and meeting with faculty advisor functionality
Updated the functionality for removing holds and approving meetings with the faculty advisor to update and utilize the database accordingly. Added a small method for getting students with a hold and displaying them but did not add one for students without holds sine that is listed as a task for another sprint.