Cleaned up code a little and added some nugetpackages
Cleaned up a warning for referencing nsubstitute when it wasn't used and added some packages for ef core for testing adding DB context to the desktop app. Desktop app might have to be remade in a different framework if we want to be able to reference .net framework libraries and unit tests.