Added notes desktop,  fixed bug in student apps
Added the ability to display notes from previous sessions in the desktop app, also updated the format of the notes to show advisor name, date and the notes. Also fixed a bug in the student app that occured when we added the holdID to the sessions table. Simply just added the hold id when assinging a students sessions.