Added views for cancelling meetings
Added the views for cancelling meetings to all app versions. For the student it has the correct restrictions such as not being able to cancel the day of a meeting/after the day of. Base controllers/controller methods added for the views and cancelling meeting messages added to associated views when the user hits yes but still needs to actually trie the functionality of cancelling the meeting inside the controllers.