Updated avaliability change for ease of access
Updated the wireframe for advisor setting their availability to be easier to use. Check boxes seemed more complex for users to use and keep track of so changed it to be drop down select to select the day for the time and drop down for the time slots as well.