Changed StudentsWithHolds list view in Advisor desktop app to display all of the advisors students instead.
Fixed bug where if you closed a view with the X at the top right of the window, the application would not close since other views are enables, only not visible.
Closing a form in any way that isnt intended will not cascade to the parent forms and close the application.

Changed ObtainStudentsWithHolds to ObtainStudentsOfAdvisor method in StudentDal to obtain all students irrespective of their hold status.